@@ -1,6 +1,5 @@
 {
   "name": "laravel-react",
-  "version": "v2.2.2",
   "scripts": {
     "dev": "npm run development",
     "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
@@ -18,56 +17,51 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "browser-sync": "^2.26.10",
-    "browser-sync-webpack-plugin": "^2.2.2",
-    "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^3.0.4",
-    "eslint-plugin-html": "^6.0.2",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-standard": "^4.0.1",
-    "immutable": "^4.0.0-rc.12",
-    "laravel-mix": "^5.0.4",
-    "redux-logger": "^3.0.6",
-    "remove-files-webpack-plugin": "1.1.3",
-    "resolve-url-loader": "^3.1.1",
-    "sass": "^1.26.10",
-    "sass-loader": "^8.0.2"
-  },
   "dependencies": {
-    "axios": "^0.19.2",
-    "bootstrap": "^4.5.0",
-    "classnames": "^2.2.6",
-    "draft-js": "^0.11.6",
-    "draftjs-to-html": "^0.9.1",
+    "@popperjs/core": "^2.5.4",
+    "axios": "^0.21.0",
+    "bootstrap": "^4.5.3",
+    "clsx": "^1.1.1",
     "font-awesome": "^4.7.0",
-    "history": "^4.10.1",
+    "history": "^5.0.0",
     "jquery": "^3.5.1",
-    "lodash": "^4.17.19",
-    "moment": "^2.27.0",
-    "popper.js": "^1.16.1",
+    "lodash": "^4.17.20",
+    "moment": "^2.29.1",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
+    "react": "^17.0.1",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.13.1",
-    "react-draft-wysiwyg": "^1.14.5",
+    "react-dom": "^17.0.1",
     "react-loadable": "^5.5.0",
-    "react-redux": "^7.2.0",
+    "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
-    "reactstrap": "^8.5.1",
+    "reactstrap": "^8.7.1",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
-    "ree-validate": "^3.3.1"
+    "ree-validate": "^3.3.2"
+  },
+  "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/preset-env": "^7.12.7",
+    "@babel/preset-react": "^7.12.7",
+    "babel-eslint": "^10.1.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "browser-sync": "^2.26.13",
+    "browser-sync-webpack-plugin": "^2.2.2",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.14.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-plugin-html": "^6.1.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-webpack-plugin": "^2.4.1",
+    "laravel-mix": "^5.0.9",
+    "redux-logger": "^3.0.6",
+    "remove-files-webpack-plugin": "^1.4.4",
+    "resolve-url-loader": "^3.1.2",
+    "sass": "^1.29.0",
+    "sass-loader": "^10.1.0"
   }
 }
