@@ -4,7 +4,6 @@ namespace App\Console;
 
 use App\Console\Commands\RefreshDatabase;
 use Illuminate\Console\Scheduling\Schedule;
-use App\Console\Commands\PassportConfigure;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
 
 class Kernel extends ConsoleKernel
@@ -15,14 +14,13 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        RefreshDatabase::class, // refresh database
-        PassportConfigure::class,
     ];
 
     /**
      * Define the application's command schedule.
      *
-     * @param  \Illuminate\Console\Scheduling\Schedule  $schedule
+     * @param \Illuminate\Console\Scheduling\Schedule $schedule
+     *
      * @return void
      */
     protected function schedule(Schedule $schedule)
