@@ -1,12 +1,12 @@
+window._ = require('lodash');
+
 /**
- * We'll load jQuery and the Bootstrap jQuery plugin which provides support
- * for JavaScript based Bootstrap features such as modals and tabs. This
- * code may be modified to fit the specific needs of your application.
+ * We'll load the axios HTTP library which allows us to easily issue requests
+ * to our Laravel back-end. This library automatically handles sending the
+ * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-try {
-  window.$ = window.jQuery = require('jquery');
-// eslint-disable-next-line no-empty
-  require('popper.js')
-  require('bootstrap')
-} catch(e) {}
+window.axios = require('axios');
+
+window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
+
