@@ -5,11 +5,11 @@ export default [
   {
     path: '/login',
     exact: true,
-    component: lazy(() => import('./pages/login/index')),
+    component: lazy(() => import('./pages/login')),
   },
   {
     path: '/register',
     exact: true,
-    component: lazy(() => import('./pages/register/index')),
+    component: lazy(() => import('./pages/register')),
   },
 ]
