@@ -6,6 +6,6 @@ export default [
     path: '/users/:id/edit',
     exact: true,
     auth: true,
-    component: lazy(() => import('./pages/edit/index')),
+    component: lazy(() => import('./pages/edit')),
   },
 ]
