@@ -3,7 +3,7 @@ import PropTypes from "prop-types"
 
 // import components
 import Header from "./components/Header"
-import Articles from "../../../../common/articles/listing/index"
+import Articles from "../../../../common/articles/listing"
 
 // import services
 import { articleListRequest } from '../../../article/service'
